+++ src/clj/rems/api/schema.clj
@@ -78,6 +78,7 @@
    :resource/ext-id s/Str
    :catalogue-item/id s/Int
    :catalogue-item/title LocalizedString
+   :catalogue-item/infourl LocalizedString
    :catalogue-item/start DateTime
    :catalogue-item/end (s/maybe DateTime)
    :catalogue-item/enabled s/Bool
@@ -160,6 +161,13 @@
   {:application/id s/Int
    :application/external-id s/Str
    :application/state s/Keyword
+   :application/todo (s/maybe (s/enum :new-application
+                                      :no-pending-requests
+                                      :resubmitted-application
+                                      :waiting-for-decision
+                                      :waiting-for-review
+                                      :waiting-for-your-decision
+                                      :waiting-for-your-review))
    :application/created DateTime
    :application/modified DateTime
    (s/optional-key :application/first-submitted) DateTime